# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetMirror content API — movies, TV series, live TV, and scraper control
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContentItemType = typeof ContentItemType[keyof typeof ContentItemType];


export const ContentItemType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export type ContentItemDubsItem = {
  subjectId: string;
  lanName: string;
  lanCode: string;
  detailPath: string;
};

export interface ContentItem {
  id: number;
  /** @nullable */
  tmdbId?: number | null;
  /** @nullable */
  subjectId?: string | null;
  type: ContentItemType;
  title: string;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  hueDark?: string | null;
  /** @nullable */
  detailPath?: string | null;
  categories?: string[];
  /** @nullable */
  qualities?: string[] | null;
  /** @nullable */
  scrapedOk?: boolean | null;
  /** @nullable */
  dubs?: ContentItemDubsItem[] | null;
}

export type ContentListResponsePagination = {
  page: number;
  limit: number;
  total: number;
};

export interface ContentListResponse {
  data: ContentItem[];
  pagination: ContentListResponsePagination;
}

export interface CategoriesResponse {
  categories: string[];
}

export type RemoteContentItemType = typeof RemoteContentItemType[keyof typeof RemoteContentItemType];


export const RemoteContentItemType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface RemoteContentItem {
  subjectId: string;
  type: RemoteContentItemType;
  title: string;
  /** @nullable */
  year?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  detailPath?: string | null;
}

export interface SearchResponse {
  local: ContentItem[];
  remote: RemoteContentItem[];
}

export interface VideoSource {
  id: number;
  contentId: number;
  /** @nullable */
  episodeId?: number | null;
  server: number;
  quality: string;
  url: string;
  /** @nullable */
  expiresAt?: string | null;
}

export interface Episode {
  id: number;
  contentId: number;
  seasonNum: number;
  episodeNum: number;
  /** @nullable */
  cdnContentId?: string | null;
  /** @nullable */
  qualities?: string[] | null;
}

export interface ContentDetailResponse {
  content: ContentItem;
  videoSources?: VideoSource[];
  episodes?: Episode[];
  /** @nullable */
  live?: boolean | null;
}

export interface QualityOption {
  quality: string;
  url: string;
  server: number;
}

export interface Caption {
  lang: string;
  name: string;
  url: string;
}

export interface VideoStream {
  url: string;
  resolution: number;
  /** @nullable */
  size?: number | null;
}

export interface VideoUrlResponse {
  url: string;
  quality: string;
  /** @nullable */
  server?: number | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  contentId?: number | null;
  /** @nullable */
  allQualities?: QualityOption[] | null;
  /** @nullable */
  streams?: VideoStream[] | null;
  /** @nullable */
  captions?: Caption[] | null;
  /** @nullable */
  live?: boolean | null;
}

export type EpisodesResponseSeasons = {[key: string]: Episode[]};

export interface EpisodesResponse {
  title: string;
  seasons: EpisodesResponseSeasons;
  totalEpisodes: number;
}

export interface LogEntry {
  ts: string;
  msg: string;
}

export type ScraperJobStatus = typeof ScraperJobStatus[keyof typeof ScraperJobStatus];


export const ScraperJobStatus = {
  pending: 'pending',
  running: 'running',
  done: 'done',
  failed: 'failed',
} as const;

export interface ScraperJob {
  id: number;
  status: ScraperJobStatus;
  jobType: string;
  /** @nullable */
  totalItems?: number | null;
  /** @nullable */
  processedItems?: number | null;
  /** @nullable */
  failedItems?: number | null;
  /** @nullable */
  currentItem?: string | null;
  /** @nullable */
  log?: LogEntry[] | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface ScraperStatus {
  isRunning: boolean;
  /** @nullable */
  activeJobId?: number | null;
  recentJobs: ScraperJob[];
}

export interface ScraperJobResponse {
  job: ScraperJob;
}

export type ScraperStartInputType = typeof ScraperStartInputType[keyof typeof ScraperStartInputType];


export const ScraperStartInputType = {
  full: 'full',
  metadata: 'metadata',
  video: 'video',
  update_check: 'update_check',
  live_tv: 'live_tv',
} as const;

export interface ScraperStartInput {
  type: ScraperStartInputType;
}

export interface ScraperStartResponse {
  jobId: number;
  message: string;
  statusUrl: string;
}

export interface LiveChannel {
  id: number;
  channelId: string;
  name: string;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  logo?: string | null;
  streamUrl: string;
  /** @nullable */
  licenseType?: string | null;
  /** @nullable */
  licenseKey?: string | null;
  /** @nullable */
  isUp?: boolean | null;
}

export interface LiveChannelsResponse {
  channels: LiveChannel[];
}

export interface LiveGroupsResponse {
  groups: string[];
}

export type ListContentParams = {
type?: ListContentType;
category?: string;
q?: string;
page?: number;
limit?: number;
};

export type ListContentType = typeof ListContentType[keyof typeof ListContentType];


export const ListContentType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export type SearchContentParams = {
/**
 * @minLength 2
 */
q: string;
};

export type GetVideoUrlParams = {
quality?: GetVideoUrlQuality;
season?: number;
episode?: number;
};

export type GetVideoUrlQuality = typeof GetVideoUrlQuality[keyof typeof GetVideoUrlQuality];


export const GetVideoUrlQuality = {
  NUMBER_480: '480',
  NUMBER_720: '720',
  NUMBER_1080: '1080',
} as const;

export type StopScraper200 = {
  ok: boolean;
};

export type ListLiveChannelsParams = {
group?: string;
};

export type RefreshLiveChannels200 = {
  ok: boolean;
  count: number;
};