# lib/api-zod/src/generated/types/categoriesResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetMirror content API — movies, TV series, live TV, and scraper control
 * OpenAPI spec version: 0.1.0
 */

export interface CategoriesResponse {
  categories: string[];
}