# lib/api-zod/src/generated/types/contentDetailResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * NetMirror content API — movies, TV series, live TV, and scraper control
 * OpenAPI spec version: 0.1.0
 */
import type { ContentItem } from './contentItem';
import type { Episode } from './episode';
import type { VideoSource } from './videoSource';

export interface ContentDetailResponse {
  content: ContentItem;
  videoSources?: VideoSource[];
  episodes?: Episode[];
  /** @nullable */
  live?: boolean | null;
}